requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
openpyxl>=3.1.2
click>=8.1.7
rich>=13.7.0
python-evtx>=0.7.4
